create schema_matching.h
