+++ torch/csrc/jit/ir.cpp
@@ -5,7 +5,7 @@
 #include <torch/csrc/autograd/function.h>
 #include <torch/csrc/jit/constants.h>
 #include <torch/csrc/jit/assertions.h>
-#include <torch/csrc/jit/script/compiler.h>
+#include <torch/csrc/jit/script/schema_matching.h>
 #include <torch/csrc/jit/passes/python_print.h>
 #include <torch/csrc/jit/passes/alias_analysis.h>
 
@@ -1475,6 +1475,15 @@ void Graph::freeBlock(Block * b) {
   all_blocks.erase(it);
 }
 
+at::ArrayRef<Value*> createTupleUnpack(Value* v) {
+  // small peephole optimization to ensure IntList attributes can still turn
+  // into constants e.g. in x.expand([3, 4])
+  if(v->node()->kind() == prim::TupleConstruct)
+    return v->node()->inputs();
+  auto & g = *v->owningGraph();
+  return g.insertNode(g.createTupleUnpack(v))->outputs();
+}
+
 PythonOp* defaultAllocPythonOp(Graph*g) {
   throw std::runtime_error("Trying to allocate a Python object without python bindings loaded");
 }
@@ -1487,5 +1496,6 @@ PythonOp* allocPythonOp(Graph* g) {
 void setAllocPythonOp(PythonOp* (*v)(Graph* g)) {
   alloc_python_op.store(v);
 }
+
 
 }} // namespace torch::jit
