+++ torch/csrc/jit/passes/to_batch.cpp
@@ -23,7 +23,7 @@ std::shared_ptr<Graph> ToBatch::getBatchOperator(const std::string& name, int64_
 std::vector<Value*> inlineUnpackedCallTo(Graph& g, Graph& callee, ArrayRef<Value*> inputs) {
   auto outputs = script::inlineCallTo(g, callee, inputs);
   if (callee.outputs().size() == 1 && callee.outputs().at(0)->type()->kind() == TupleType::Kind) {
-    auto tc = script::createTupleUnpack(outputs.at(0));
+    auto tc = createTupleUnpack(outputs.at(0));
     outputs = std::vector<Value*>(tc.begin(), tc.end());
   }
   return outputs;
@@ -527,7 +527,7 @@ std::shared_ptr<Graph> to_batch_graph(std::shared_ptr<Graph> graph) {
   // lower the tuple before the pass
   if (graph->outputs().at(0)->type()->kind() == TupleType::Kind) {
     graph = graph->copy();
-    auto outs = script::createTupleUnpack(graph->outputs().at(0));
+    auto outs = createTupleUnpack(graph->outputs().at(0));
     graph->eraseOutput(0);
     for(auto o : outs)
       graph->registerOutput(o);
